requirements.txt
tk==0.1.0
absl-py==2.1.0
accelerate==1.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_record==0.5.1
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
bokeh==2.0.1
catboost==1.2.7
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
cons==0.4.6
contourpy==1.3.0
cryptography==3.4.8
cycler==0.12.1
dacite==1.8.1
dbus-python==1.2.18
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
dm-tree==0.1.8
docstring_parser==0.16
etils==1.10.0
etuples==0.3.9
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
fsspec==2024.10.0
gast==0.6.0
GDAL==3.4.1
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
htmlmin==0.1.12
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
ImageHash==4.3.1
imageio==2.36.0
immutabledict==4.2.0
importlib-metadata==4.6.4
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.4.1
keyring==23.5.0
kiwisolver==1.4.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
logical-unification==0.4.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
miniKanren==1.0.3
mistune==3.0.2
ml-dtypes==0.4.1
more-itertools==8.10.0
mpmath==1.3.0
multimethod==1.12
multipledispatch==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.0
opt_einsum==3.4.0
optree==0.13.0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
phik==0.12.4
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
prometheus_client==0.21.0
promise==2.3
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==3.2.0
pytensor==2.25.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyWavelets==1.7.0
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.3
rpds-py==0.20.0
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
simple-parsing==0.1.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-datasets==4.9.6
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
tensorflow-probability==0.24.0
tensorstore==0.1.66
termcolor==2.5.0
terminado==0.18.1
tf-slim==1.1.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.4.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.2
toolz==1.0.0
torch @ https://download.pytorch.org/whl/cu121_full/torch-2.5.0%2Bcu121-cp310-cp310-linux_x86_64.whl#sha256=860618b5d6ffd964c840e3fba6a911489ca359826031f35a5b7bac6d74ee65cb
torchaudio @ https://download.pytorch.org/whl/cu121_full/torchaudio-2.5.0%2Bcu121-cp310-cp310-linux_x86_64.whl#sha256=53012f530ffc9023b4416a56ba7ee02f486bc484ce6037df40927f0f87aa1ef4
torchsummary==1.5.1
torchvision @ https://download.pytorch.org/whl/cu121_full/torchvision-0.20.0%2Bcu121-cp310-cp310-linux_x86_64.whl#sha256=7a1023c04ad2a77e8350a992a025d69e03214230390f2cf2c32b1e8e8beba7e9
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
transformers==4.44.2
typeguard==4.4.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
visions==0.7.6
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wordcloud==1.9.3
wrapt==1.16.0
ydata-profiling==4.12.0
zipp==3.20.2
